+++ src/InfinitesimalGenerators.jl
@@ -1,11 +1,10 @@
 module InfinitesimalGenerators
 
 using Distributions: Normal, Gamma, quantile
-using FillArrays: Ones, Zeros
+using FillArrays: Ones
 using LinearAlgebra: LinearAlgebra, Adjoint, Diagonal, LAPACKException, SingularException, Transpose, Tridiagonal, I, Symmetric, diag, eigmin, factorize, kron, ldiv!
 using SparseArrays: blockdiag, findnz, sparse, spzeros
 using Roots: fzero
-using BlockBandedMatrices: BandedBlockBandedMatrix, Block
 
 include("derivatives.jl")
 include("matrix_operators/check_generator.jl")